+++ OOP - Initial Prototype/Game.h
@@ -6,12 +6,12 @@
 class Game
 {
 	public:
-		Game(int ,const std::string&, const std::string&, float, int);
+		Game(int ,const std::string&, const std::string&, int, int);
 		~Game();
 
 		const std::string& GetName() const;
 		const std::string& GetDescription() const;
-		float GetCost() const;
+		int GetCost() const;
 		int getId();
 		int getAgeRating();
